% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctams.R
\docType{data}
\name{ctams}
\alias{ctams}
\title{Ctams Database}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{Node}{Drug and cell line conditions}
  \item{m}{Number of edges atributed to node}
  \item{edges}{list of edges seperated by ;}
}
}
\usage{
ctams
}
\description{
Ctams Database
}
\keyword{datasets}
