% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drug_marker_enrichment.R
\name{Drug.Marker.Enrichment}
\alias{Drug.Marker.Enrichment}
\title{Drug marker Enrichment}
\usage{
Drug.Marker.Enrichment(df.ppindex = dat, marker_db = df.aml)
}
\arguments{
\item{df.ppIndex}{df.ppindex is phosphoproteomics, proteomics or transcriptomics data with rownames being identifiers}

\item{marker_db}{marker database to use to get enrichment}
}
\description{
Drug marker Enrichment
}
