% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugMarkerEnrichment.R
\name{DrugMarkerEnrichment}
\alias{DrugMarkerEnrichment}
\title{A score of marker ratios}
\usage{
MarkerZscores(df, marker_database = "aml.markers", scale = T)
}
\arguments{
\item{df}{dataframe of peptide intensities cant be used for DMSO controls yet}

\item{scale}{if TRUE data will be scaled before putting in ksea}

\item{marker_database}{markers database to be used}

\item{input}{datatype input can be phospho or prot}
}
\description{
Get zscores between sensitive and resistant markers
}
