% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterSensitivity.R
\name{FilterSensitivity}
\alias{FilterSensitivity}
\title{Filter Drugs by sensitivity data}
\usage{
FilterSensitivity(
  sensitivity_database = NULL,
  cell_annotations = NULL,
  cell_lines = NULL,
  tissue_type = c("haematopoietic_and_lymphoid_tissue", "liver", "oesophagus"),
  iqr_tolerance = 0.04,
  n_coverage = 0.8
)
}
\arguments{
\item{sensitivity_database}{sensitvitiy data with cell lines as rownamed and drugs as columns}

\item{cell_annotations}{dataframe with cell line annotations cell names must be passed through make.names function and placed in "cell.name" column. Tumour type must be stored in "unique.tissue" column.}

\item{cell_lines}{Cell lines present in omics data}

\item{tissue_type}{Tissue types to filter sensitvity by the default is those used in DRUML panel ("haematopoietic_and_lymphoid_tissue", "liver", "oesophagus")}

\item{iqr_tolerance}{IQR range cut off to filter drug list}

\item{n_coverage}{n_coverage cut off required to ensure models are reliable}
}
\description{
Filter Drugs by sensitivity data
}
