% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIntDB.R
\name{GetIntMarkerDB}
\alias{GetIntMarkerDB}
\title{A score of marker ratios}
\usage{
GetIntMarkerDB("aml_prot_markers")
}
\arguments{
\item{database}{database to draw into environment}
}
\description{
draw in internal databases
}
