% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimiseMarkers.R
\name{OptimiseMarkers}
\alias{OptimiseMarkers}
\title{Optimise marker set for machine learning}
\usage{
OptimiseMarkers(EMSRoutput = EMSRout, drug = "ABT-119", df_response = df.aac, df_input = df.phos)
}
\arguments{
\item{EMSRoutput}{output of GenerateEMSR function (get_limma must be False)}

\item{drug}{Drug which the markers will be generated for}

\item{df_response}{Dataframe with rownames as cell lines and drug responses}

\item{df_input}{= Input dataset, rownames must be variables}

\item{nfolds}{number of points of analysis to do to optimise marker lenghts}

\item{maxmarker_res}{The maximum number of resistant markers to be considered. If left NULL the total number of resistant markers will be used.}

\item{maxmarker_sens}{The maximum number of sensitive markers to be considered. If left NULL the total number of sensitive markers will be used.}

\item{var1}{Variable to use for input variable sorting can be  n_pvalue_cutoff, n_increase, n_decrease, ratio, n_increase_pfilt, n_decrease_pfilt, ratio_pfilt, mean_fold, mean_fold_pfilt, mean_}

\item{var2}{= Second variable to sort input variables}

\item{sens_class}{variable to use to classify variables as a marker of sensitivity or resistance (require negative values) e.g mean_fold and ratio}

\item{computational_load}{proportion of cores to use for analysis}

\item{graph}{if True a graph of pvalue spread across marker combinations will be generated}
}
\description{
Optimise marker set for machine learning
}
