% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRUMLPredict.R
\name{predict.drug.sensitivity.based.on.distance.markers}
\alias{predict.drug.sensitivity.based.on.distance.markers}
\title{Make a heatmap of marker enrichment}
\usage{
predict.drug.sensitivity.based.on.distance.markers(markers = GetMarkers("barasetib")$sensitive$.,inputdata = LoadData("PDBphos"),tissuefilter = "haematopoietic_and_lymphoid_tissue",metric = "aac", drug = "ABT-199")
}
\arguments{
\item{rmse.cutoff}{error cut off to define models as reliable}

\item{solid.or.aml}{which models to predict sensitvity}

\item{predict.drug.sensitivity.based.on.distance.markers}{List of markers to filter the input data by}

\item{dataframe}{to use for the heatmap valuesdf.ppindex is phosphoproteomics, proteomics or transcriptomics data with rownames being identifiers}

\item{type}{"phospho","protein","RNA","ksea edges","reactome prot"}
}
\description{
Make a heatmap of marker enrichment
}
