% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRUMLDataCompat.R
\name{TestExternalDataComp}
\alias{TestExternalDataComp}
\title{Recaluclate Error paramaters for external data to account for missing variables}
\usage{
TestExternalDataComp(markers = GetMarkers("barasetib")$sensitive$.,inputdata = LoadData("PDBphos"),tissuefilter = "haematopoietic_and_lymphoid_tissue",metric = "aac", drug = "ABT-199")
}
\arguments{
\item{TestExternalDataComp}{List of markers to filter the input data by}

\item{dataframe}{to use for the heatmap valuesdf.ppindex is phosphoproteomics, proteomics or transcriptomics data with rownames being identifiers}

\item{rmse.cutoff}{error cut off to define models as reliable}

\item{solid.or.aml}{which models to predict sensitvity}

\item{type}{"phospho","protein","RNA","ksea edges","reactome prot"}
}
\description{
Recaluclate Error paramaters for external data to account for missing variables
}
