% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_databases.R
\name{Update_databases}
\alias{Update_databases}
\alias{UpdateDatabases}
\title{Update Internal Databases for KSEA}
\usage{
Update_databases()
}
\description{
Update databases from databases files folder
}
