% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateEMSR.R
\name{GenerateEMSR}
\alias{GenerateEMSR}
\title{Emperical Markers of Sensitivity and resistance}
\usage{
GenerateEMSR(
  df_input = NULL,
  df_response = NULL,
  drug = "barasertib",
  resampling.times = 10,
  p.cutoff = 0.05,
  direction_percentage = 0.8,
  fold.cut.off = 0.8,
  marker.n.cutoff = 0.3,
  save_xlsx = F,
  save_name = NULL,
  computational_load = NULL,
  pfilt = T
)
}
\arguments{
\item{df_input}{input data with cell lines as colnames and varaibles as rownames defaults to our dataset}

\item{df_response}{response values dataframe with colnames as cell lines and drugs as column. defaults to our dataset}

\item{drug}{drug generate markers for}

\item{computational_load}{decimal fraction of cores you want to use. If left as NULL only 1 core will be used.}
}
\description{
Emperical Markers of Sensitivity and resistance
}
